Reorganized project folder
