+++ .vscode/launch.json
@@ -5,6 +5,17 @@
     "version": "0.2.0",
     "configurations": [
         {
+            "name": "Launch Package",
+            "type": "go",
+            "request": "launch",
+            "mode": "debug",
+            "program": "${workspaceFolder}",
+            "args": [
+                "-gallery",
+                "vangroan.deviantart.com"
+            ]
+        },
+        {
             "name": "Run Tests",
             "type": "go",
             "request": "launch",
