+++ scrapers/base.go
@@ -1,11 +1,10 @@
 package scrapers
 
 import (
-	artdl "github.com/vangroan/art-dl"
+	artdl "github.com/vangroan/art-dl/common"
 )
 
 // baseScraper contains useful, commonly used operations.
 type baseScraper struct {
 	config *artdl.Config
 }
-
